backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
shapely
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
